Old tasks no longer get deleted when moved into the future
Old tasks which are not completed will not be deleted when moved 2 days into the furute.
Tasks which are completed, however, did not change behaviour and will get deleted

+++ src/components/overview/Overview.tsx
@@ -51,7 +51,9 @@ class Overview extends React.Component<Props, { tasks: [], error: boolean, autht
     }
   };
 
-  createNewTaskFromOldTask = async (oldId: string, day: string, chore: string, pic: string, blame: string) => {
+  createNewTaskFromOldTask =
+      async (oldId: string, day: string, chore: string, pic: string, blame: string, done: boolean) => {
+
     const { authtoken } = this.state;
     const dayForNewDocument = DayUtil.getDayTwoDaysFromDay(day);
 
@@ -64,7 +66,10 @@ class Overview extends React.Component<Props, { tasks: [], error: boolean, autht
     };
 
     await Api.addDocument(newDocument, authtoken);
-    await Api.deleteDocument(oldId, authtoken);
+
+    if (done) {
+      await Api.deleteDocument(oldId, authtoken);
+    }
   };
 
   render() {
